major rehaul of optimize. Batching is now using pytorch Dataset and DataLoader
