+++ optimize/example_run.py
@@ -1,27 +1,63 @@
 #!/usr/bin/env python3
 
-from ParamFitter import ParamFitter
 import argparse
+import sys
+import traceback
+from torch.utils.data import DataLoader
 
-def main(param_list, input_file, lr, epochs, seed):
-    param_fit = ParamFitter(param_list, lr=lr) 
-    param_fit.make_target_sim(seed=seed)
-    param_fit.load_data(filename=input_file)
-    param_fit.fit(epochs=epochs)
+from .fit_params import ParamFitter
+from .dataio import TracksDataset
 
+def main(config):
+    dataset = TracksDataset(filename=config.input_file)
+    tracks_dataloader = DataLoader(dataset,
+                                  shuffle=True, batch_size=config.batch_sz,
+                                  pin_memory=True, num_workers=config.num_workers)
+    param_fit = ParamFitter(config.param_list, dataset.get_track_fields(),
+                            track_chunk=config.track_chunk, pixel_chunk=config.pixel_chunk,
+                            detector_props=config.detector_props, pixel_layouts=config.pixel_layouts,
+                            load_checkpoint=config.load_checkpoint, lr=config.lr)
+    param_fit.make_target_sim(seed=config.seed)
+    param_fit.fit(tracks_dataloader, epochs=config.epochs)
 
-if __name__ == "__main__":
+    return 0, 'Fitting successful'
+
+if __name__ == '__main__':
     parser = argparse.ArgumentParser()
-    parser.add_argument("--params", dest="param_list", default=[], nargs="+", 
+    parser.add_argument("--params", dest="param_list", default=[], nargs="+", required=True,
                         help="List of parameters to optimize. See consts_ep.py")
-    parser.add_argument("--input-file", dest="input_file", default="",
+    parser.add_argument("--input_file", dest="input_file",
+                        default="/sdf/group/neutrino/cyifan/muon-sim/fake_data_S1/edepsim-output.h5",
                         help="Input data file")
+    parser.add_argument("--detector_props", dest="detector_props",
+                        default="larndsim/detector_properties/module0.yaml",
+                        help="Path to detector properties YAML file")
+    parser.add_argument("--pixel_layouts", dest="pixel_layouts",
+                        default="larndsim/pixel_layouts/multi_tile_layout-2.2.16.yaml",
+                        help="Path to pixel layouts YAML file")
+    parser.add_argument("--load_checkpoint", dest="load_checkpoint", type=str, default=None,
+                        help="Path to checkpoint Pickle (pkl) file")
+    parser.add_argument("--track_chunk", dest="track_chunk", default=1, type=int,
+                        help="Track chunk size used in simulation.")
+    parser.add_argument("--pixel_chunk", dest="pixel_chunk", default=1, type=int,
+                        help="Pixel chunk size used in simulation.")
+    parser.add_argument('--num_workers', type=int, default=4,
+                        help='The number of worker threads to use for the dataloader.')
     parser.add_argument("--lr", dest="lr", default=1e1, type=float,
                         help="Learning rate -- used for all params")
+    parser.add_argument("--batch_sz", dest="batch_sz", default=2, type=int,
+                        help="Batch size for fitting (tracks).")
     parser.add_argument("--epochs", dest="epochs", default=100, type=int,
                         help="Number of epochs")
     parser.add_argument("--seed", dest="seed", default=2, type=int,
                         help="Random seed for target construction")
-   
-    args = parser.parse_args()
-    main(args.param_list, args.input_file, args.lr, args.epochs, args.seed)
+    try:
+        args = parser.parse_args()
+        retval, status_message = main(args)
+    except Exception as e:
+        print(traceback.format_exc(), file=sys.stderr)
+        retval = 1
+        status_message = 'Error: Fitting failed.'
+
+    print(status_message)
+    exit(retval)
