+++ optimize/utils.py
@@ -51,8 +51,9 @@ def all_sim(sim, selected_tracks, fields, event_id_map, unique_eventIDs, return_
                                           max_length_torch,
                                           fields=fields)
 
-    unique_pix_torch = torch.empty((0, 2))
-    pixels_signals_torch = torch.zeros((len(unique_pix_torch), len(sim.time_ticks)*50))
+    unique_pix_torch = torch.empty((0, 2), device=neighboring_pixels_torch.device)
+    pixels_signals_torch = torch.zeros((len(unique_pix_torch), len(sim.time_ticks)*50),
+                                       device=unique_pix_torch.device)
 
     shapes_torch = neighboring_pixels_torch.shape
     joined_torch = neighboring_pixels_torch.reshape(shapes_torch[0]*shapes_torch[1], 2)
@@ -61,10 +62,12 @@ def all_sim(sim, selected_tracks, fields, event_id_map, unique_eventIDs, return_
     this_unique_pix_torch = this_unique_pix_torch[(this_unique_pix_torch[:,0] != -1) & (this_unique_pix_torch[:,1] != -1),:]
     unique_pix_torch = torch.cat((unique_pix_torch, this_unique_pix_torch),dim=0)
 
-    this_pixels_signals_torch = torch.zeros((len(this_unique_pix_torch), len(sim.time_ticks)*50))
+    this_pixels_signals_torch = torch.zeros((len(this_unique_pix_torch), len(sim.time_ticks)*50),
+                                            device=unique_pix_torch.device)
     pixels_signals_torch = torch.cat((pixels_signals_torch, this_pixels_signals_torch), dim=0)
 
-    pixel_index_map_torch = torch.full((selected_tracks.shape[0], neighboring_pixels_torch.shape[1]), -1)
+    pixel_index_map_torch = torch.full((selected_tracks.shape[0], neighboring_pixels_torch.shape[1]), -1,
+                                       device=unique_pix_torch.device)
     compare_torch = (neighboring_pixels_torch[..., np.newaxis, :] == unique_pix_torch)
 
     indices_torch = torch.where(torch.logical_and(compare_torch[..., 0], compare_torch[...,1]))
@@ -95,7 +98,7 @@ def update_grad_param(sim, name, value):
 # ADC counts given as list of pixels. Better for loss to embed this in the "full" pixel space
 def embed_adc_list(sim, adc_list, unique_pix):
     zero_val = sim.digitize(torch.tensor(0)).item()
-    new_list = torch.ones((sim.n_pixels[0], sim.n_pixels[1], adc_list.shape[1]))*zero_val
+    new_list = torch.ones((sim.n_pixels[0], sim.n_pixels[1], adc_list.shape[1]), device=unique_pix.device)*zero_val
 
     plane_id = unique_pix[..., 0] // sim.n_pixels[0]
     unique_pix[..., 0] = unique_pix[..., 0] - sim.n_pixels[0] * plane_id
